context: Ensure context name is valid on import
(cherry picked from commit 9ecc69d17e40cd0a03b6b03d8421bc415936d231)
